+++ frontend/src/stores/NotesByBookStore.js
@@ -12,42 +12,45 @@ class NotesByBookStore {
     }
 
     async getAll(bookID) {
-        let response = await fetch(SERVER_URL + `/books/${bookID}/notes`);
-        let data = await response.json();
-        
-        this.notes = data;
-        this.emitter.emit(CODES.CODE_GET_ALL_NOTES_BY_BOOKS);
+        if (bookID > 0) {
+            let response = await fetch(SERVER_URL + `/books/${bookID}/notes`);
+            let data = await response.json();
+            
+            this.notes = data;
+            this.emitter.emit(CODES.CODE_GET_ALL_NOTES_BY_BOOKS);
+        }
     }
 
     async getById(id) {
-        let response = await fetch(SERVER_URL + `/notes/${id}`);
-        let data = await response.json();
+        if (id > 0) {
+            let response = await fetch(SERVER_URL + `/notes/${id}`);
+            let data = await response.json();
 
-        this.object = data;
-        this.emitter.emit(CODES.CODE_GET_NOTE_BY_ID);
-
-    }
-
-    async create(note) {
-
+            this.object = data;
+            this.emitter.emit(CODES.CODE_GET_NOTE_BY_ID);
+        }
     }
 
     async update(id, note) {
-        await fetch(SERVER_URL + `/notes/${id}`, {
-            method: 'PUT', 
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(note)
-        });
+        if (id > 0) {
+            await fetch(SERVER_URL + `/notes/${id}`, {
+                method: 'PUT', 
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(note)
+            });
+        }
     }
 
     async delete(id) {
-        await fetch(SERVER_URL + `/notes/${id}`, {
-            method: 'DELETE', 
-        });
-    
-        this.getAll();
+        if (id > 0) {
+            await fetch(SERVER_URL + `/notes/${id}`, {
+                method: 'DELETE', 
+            });
+        
+            this.getAll();
+        }
     }
 }
