+++ frontend/src/stores/NotesStore.js
@@ -21,12 +21,13 @@ class NotesStore {
     }
 
     async getById(id) {
+        if (id > 0) {
+            let response = await fetch(SERVER_URL + `/notes/${id}`);
+            let data = await response.json();
 
-        let response = await fetch(SERVER_URL + `/notes/${id}`);
-        let data = await response.json();
-
-        this.object = data;
-        this.emitter.emit(CODES.CODE_GET_NOTE_BY_ID);
+            this.object = data;
+            this.emitter.emit(CODES.CODE_GET_NOTE_BY_ID);
+        }
     }
 
     async create(note) {
@@ -44,20 +45,24 @@ class NotesStore {
     }
 
     async update(id, note) {
-        await fetch(SERVER_URL + `/notes/${id}`, {
-            method: 'PUT', 
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(note)
-        });
+        if (id > 0) {
+            await fetch(SERVER_URL + `/notes/${id}`, {
+                method: 'PUT', 
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(note)
+            });
+        }
     }
 
     async delete(id) {
-        await fetch(SERVER_URL + `/notes/${id}`, {
-            method: 'DELETE', 
-        });
-        this.getAll();
+        if (id > 0) {
+            await fetch(SERVER_URL + `/notes/${id}`, {
+                method: 'DELETE', 
+            });
+            this.getAll();
+        }
     }
 }
